Pull FieldAndFormat into its own top-level class.
